fix highlighting and tour progression issues
