+++ example-apps/react-js/src/components/ConfigPanel.jsx
@@ -2,6 +2,10 @@ import { useState } from "react";
 import PropTypes from "prop-types";
 import { RotateCcw, X, ChevronDown, ChevronUp } from "lucide-react";
 import styles from "./ConfigPanel.module.css";
+import { createTour } from "@utils/onboardingTour";
+import { TOUR_KEY } from "@utils/onboardingTour";
+import { driver } from "driver.js";
+import "driver.js/dist/driver.css";
 
 const ConfigPanel = ({
   numColors,
@@ -22,6 +26,12 @@ const ConfigPanel = ({
   isProcessing = false,
   className = "",
 }) => {
+  {/* Continued onboarding tour set up */}
+  const savedStep = localStorage.getItem(TOUR_KEY);
+  if (savedStep !== null) {
+    createTour().drive(Number(savedStep));
+  }
+
   const [isAdvancedOpen, setIsAdvancedOpen] = useState(false);
 
   return (
